+++ tests/models/job/test_orders.py
@@ -287,7 +287,7 @@ class TestOrderModel(BaseModelTest):
                 'first_name': 'J.  ',
                 'last_name': 'Hartman-Zwiers',
                 'order_id': str(order.id),
-                'mobile': '+33 31356240391',
+                'mobile': '+4917635573242',
                 'email': 'info@sdlyonne.com',
                 'timezone': 'Europe/Berlin',
                 'formatted_address': 'Nansenstraße 17, 12047 Berlin, Germany',
@@ -819,7 +819,7 @@ class TestOrderModel(BaseModelTest):
 
         assert assignment.state == 'completed'
         assert assignment.state_history[-1]['transition'] == 'complete'
-        assert assignment.state_history[-1]['message'] is None
+        assert assignment.state_history[-1]['message'] == ''
 
     @pytest.mark.parametrize('origin_state', ['in_qa', 'refused', ])
     @pytest.mark.parametrize('ass_origin_state', ['in_qa', 'refused', 'approved', ])