+++ HISTORY.rst
@@ -52,10 +52,15 @@ History
     * Refactor dashboards: PM, Customer, Professional, Bizdev, QA, Scout, Finance (rudaporto).
     * Added Order.current_type as a field in the Assignment model and listing and as a filter in the assignments view (rudaporto).
 
-2.1.35 (Unreleased)
+2.1.36 (2017-08-06)
 -------------------
 
-    * TODO
+    * Card #519: Return profile information when billing info is not available (ericof).
+
+2.1.35 (2017-08-04)
+-------------------
+
+    * Card #519: Add first_name and last_name columns to professional export (ericof).
 
 2.1.34 (2017-07-31)
 -------------------